Use TxMerge for smuggle import (#4014)
* Remove Raven.Smuggler for now as it use the old version of dotnet/dnx

* Add cancellation token

* Use TxMerger for smuggler import of documents.
Import identities only if the value is bigger than existing one.

* Refactor smuggler import to DatabaseDataImporter

* Handle warning in smuggler import

* Avoid the call for /replication/topology on the root url which isn't supported anymore

* Avoid an error that database is not exist

* We should handler to first read object differently, and not skip it as unrecognized.

* Use dbSuffix in order to preserve the test name in the database name

* Honor the Limit in DatabaseDataExporter

* Send the Limit in smuggler options to the server

+++ src/Raven.Client/Smuggler/DatabaseSmugglerOptions.cs
@@ -1,12 +1,3 @@
-using System;
-using System.Collections.Generic;
-using Raven.Abstractions.Data;
-using Raven.Abstractions.Json;
-using Raven.Imports.Newtonsoft.Json;
-using Raven.Imports.Newtonsoft.Json.Linq;
-using Raven.Json.Linq;
-using System.Linq;
-
 namespace Raven.Client.Smuggler
 {
     public class DatabaseSmugglerOptions
@@ -22,7 +13,7 @@ namespace Raven.Client.Smuggler
 
         public bool IgnoreErrorsAndContinue { get; set; }
 
-        public long? Limit { get; set; }
+        public int? Limit { get; set; }
 
         public string TransformScript { get; set; }
